:sparkles: Implemented firebase pulling all timelines and then working with the first pulled timeline

+++ vision-quickstart/app/src/main/java/com/google/mlkit/vision/demo/java/posedetector/pointTypes/TriPointAngle.java
@@ -25,7 +25,7 @@ public class TriPointAngle implements Point {
     public TriPointAngle(JSONObject json) {
         try {
             toTrack = new ArrayList<>();
-            target = (int) json.get("angle");
+            target =  ((Double)json.get("angle")).intValue();
             leniency = Double.parseDouble(json.get("leniency").toString());
             JSONArray points = (JSONArray) json.get("toTrack");
